break out experimental options tab
- update Options docs page
- move Chat Images option here
- disable empty Images options tab
